refactor: move meta triple utils to their own file
